Moved the initialize replica set logic to the connection.
